feat: create default data entity with collection source and read source
