+++ apps/web/core/state/editor/block-types.ts
@@ -1,7 +1,7 @@
 import { SYSTEM_IDS } from '@geogenesis/sdk';
 import { INITIAL_COLLECTION_ITEM_INDEX_VALUE } from '@geogenesis/sdk/constants';
 
-import { StoreRelation } from '~/core/database/write';
+import { StoreRelation } from '~/core/database/types';
 import { EntityId } from '~/core/io/schema';
 
 type BlockTypeId = typeof SYSTEM_IDS.TEXT_BLOCK | typeof SYSTEM_IDS.IMAGE_BLOCK | typeof SYSTEM_IDS.TABLE_BLOCK;
