+++ apps/web/core/state/editor/utils.ts
@@ -1,10 +1,5 @@
-import { SYSTEM_IDS } from '@geogenesis/sdk';
-import { INITIAL_COLLECTION_ITEM_INDEX_VALUE } from '@geogenesis/sdk/constants';
 import { JSONContent, generateHTML } from '@tiptap/react';
 import pluralize from 'pluralize';
-
-import { StoreRelation } from '~/core/database/write';
-import { EntityId } from '~/core/io/schema';
 
 import { tiptapExtensions } from '~/partials/editor/extensions';
